+++ components/sql/src/test/scala/pl/touk/nussknacker/sql/service/DatabaseLookupEnricherWithCacheTest.scala
@@ -14,11 +14,7 @@ class DatabaseLookupEnricherWithCacheTest extends BaseHsqlQueryEnricherTest {
   import scala.concurrent.duration._
 
   override val service =
-    new DatabaseLookupEnricher(
-      hsqlDbPoolConfig,
-      new MetaDataProviderFactory().create(hsqlDbPoolConfig),
-      displayDbErrors = false
-    )
+    new DatabaseLookupEnricher(hsqlDbPoolConfig, new MetaDataProviderFactory().create(hsqlDbPoolConfig))
 
   override val prepareHsqlDDLs: List[String] = List(
     "CREATE TABLE persons (id INT, name VARCHAR(40));",