+++ components/sql/src/test/scala/pl/touk/nussknacker/sql/service/DatabaseQueryEnricherValidationTest.scala
@@ -14,11 +14,7 @@ import java.time.Duration
 class DatabaseQueryEnricherValidationTest extends BaseHsqlQueryEnricherTest {
 
   override val service =
-    new DatabaseQueryEnricher(
-      hsqlDbPoolConfig,
-      new MetaDataProviderFactory().create(hsqlDbPoolConfig),
-      displayDbErrors = false
-    )
+    new DatabaseQueryEnricher(hsqlDbPoolConfig, new MetaDataProviderFactory().create(hsqlDbPoolConfig))
 
   override val prepareHsqlDDLs: List[String] = List(
     "CREATE TABLE persons (id INT, name VARCHAR(40));",