+++ components/sql/src/test/scala/pl/touk/nussknacker/sql/service/DatabaseQueryEnricherWithCacheTest.scala
@@ -19,11 +19,7 @@ class DatabaseQueryEnricherWithCacheTest extends BaseHsqlQueryEnricherTest {
   )
 
   override val service =
-    new DatabaseQueryEnricher(
-      hsqlDbPoolConfig,
-      new MetaDataProviderFactory().create(hsqlDbPoolConfig),
-      displayDbErrors = false
-    )
+    new DatabaseQueryEnricher(hsqlDbPoolConfig, new MetaDataProviderFactory().create(hsqlDbPoolConfig))
 
   test("DatabaseQueryEnricher#implementation with cache") {
     val query = "select * from persons where id = ?"